+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @orderly.network/futures
 
+## 4.0.0
+
+### Major Changes
+
+- 113773f: integrate solana
+
+### Patch Changes
+
+- b033ce0: test
+- b15becb: test
+- Updated dependencies [113773f]
+- Updated dependencies [b033ce0]
+- Updated dependencies [b15becb]
+  - @orderly.network/types@2.0.0
+  - @orderly.network/utils@2.0.0
+
+## 4.0.0-alpha.2
+
+### Patch Changes
+
+- test
+- Updated dependencies
+  - @orderly.network/types@2.0.0-alpha.2
+  - @orderly.network/utils@2.0.0-alpha.2
+
+## 4.0.0-alpha.1
+
+### Patch Changes
+
+- test
+- Updated dependencies
+  - @orderly.network/types@2.0.0-alpha.1
+  - @orderly.network/utils@2.0.0-alpha.1
+
+## 4.0.0-alpha.0
+
+### Major Changes
+
+- integrate solana
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/types@2.0.0-alpha.0
+  - @orderly.network/utils@2.0.0-alpha.0
+
 ## 3.3.12
 
 ### Patch Changes